Update patch set 1
Patch Set 1:

That would work for some cases, but not in the following:

- An non-workspace file has the path "C:/foo/pom.xml"
- There is a project in the workspace called "foo" with a file called "pom.xml" (its file-system path is e.g. "C:/does/not/matter/pom.xml")

When the IResource for "C:/foo/pom.xml" is created in GitContainerMapping, the following is done:

	IFile file = ROOT.getFileForLocation(childLocation);
	if (file == null)
		file = ROOT.getFile(childLocation);

getFileForLocation would return null because it's not known in the workspace. But getFile would return the resource behind the "pom.xml" in the "foo" project, because the path argument of getFile is a workspace-path!

Later, bad things would happen because exists actually returns true for the above resource, but it would be the wrong one. Ugh.


Another idea would be this: If getFileForLocation returns null, use getFile(new Path("._egit_internal_project_for_synchronization").append(childLocation)). Later, strip this known path in stripWorkDir.

It's also a hack but as long as no project named like this exists, it would not have the above problem.

Patch-set: 1