Update patch set 1
Patch Set 1: (2 inline comments)

This can be fairly easily unit tested by querying the file history and checking that all revisions are available. What I was missing was a test fixture with which to create sample repositories with the necessary setup (multiple branches and commits).

I'll take a look at GitTestCase and see if I can manage to put tests together for this feature.

Patch-set: 1