Update patch set 2
Patch Set 2:

Added specific tests to check that the full history is properly returned whatever the position of HEAD. These tests fail when not using this patch.

I have not pulled-up the two "markStart" methods : I do not know where to put them. I copied them from GitHistoryPage, CommitSelectionDialog uses something very similar too... and I did not want to create a new "utils" class only for these methods... maybe there proper place would be on RevWalk?

As for "getTargets" : I have not tested it for three reasons :
 - It depends on the timing (see below),
 - It is never used anywhere by the platform, 
 - Its documentation is not in sync with the existing implementations.

This methods is expected to return "the direct descendants of the given revision". Such a method should be on the HistoryProvider, not on the FileHistory. CVS and local history use this method to return "all descendants of the given revision that also have diffs on the 'current' (?) file". That's is what I have implemented in this patch.

Since it uses the timestamp of the revisions to determine "all descendants" (returning only the "direct" descendants would not have the same limitation since we could make use of RevCommit.getParents()), it is dependant on the timing : in unit tests, we create all commits in the same second, so all timestamps are identical.

Patch-set: 2