Update patch set 6
Patch Set 6: I would prefer that you didn't submit this

there was a bug a while ago, that made staging hell slow, and in that case this change helped a lot. but: the bug disappeared, and staging is blazingly fast now again, even for larger numbers. i don't see the need for this change anymore, actually. abandon?

Patch-set: 6