Consider the ancestor when comparing with remote revisions
This changes all of the actions where comparing with a common ancestor
makes sense :
1. Compare With > Commit
2. Compare With > Branch, Tag, Reference
3. Double-click a commit in the history view
4. Right-click a commit in the history view > Compare With Workspace
5. Right-click a file in the history view > Compare With Workspace
6. Right-click a file in the commit viewer > Compare With Workspace

The only two actions that were previously ignoring common ancestors
were the comparison with a branch tag or reference, and comparisons
from history/commit viewer "with the workspace". All were modified
in order to factorize common code as much as possible.

Comparisons of files located out of the workspace (java.io.File) was
not functional because of a regression in HistoryPageInput introduced
by 5246f101fed6fe7541aa1ce8f9fec5b778beb788 that made the "singleFile"
input of the history view a Boolean instead of an actual File.
Furthermore, comparing these "local" files, which can be done through
actions numbered 3 to 6 above was incoherent : while 3 and 4 compared
the HEAD revision, actions 5 and 6 compared the working directory,
including local changes. All four now consistently take the local
changes into account.

Bug: 400943
Change-Id: I26bce7a1bde3b944f10f40eacbbdb7c41e486c0b