Consider the ancestor when comparing with remote revisions
This changes all of the actions where comparing with a common ancestor
makes sense :
- Compare With > Commit
- Compare With > Branch, Tag, Reference
- Double-click a commit in the history view
- Right-click a commit in the history view > Compare With Workspace
- Right-click a file in the history view > Compare With Workspace
- Right-click a file in the commit viewer > Compare With Workspace

The only two actions that were previously ignoring common ancestors
were the comparison with a branch tag or reference, and comparisons
from history/commit viewer "with the workspace". All were modified
in order to factorize common code as much as possible.

Bug: 400943
Change-Id: I26bce7a1bde3b944f10f40eacbbdb7c41e486c0b