Update patch set 2
Patch Set 2:

It's somewhat easier to move code within EGit, but it seem more appropriate to move these methods to JGit.

To move code from EGit to JGit you must ask the authors for permission due to the license change implied. Getting permissions is not likely to be a problem.

Make the refactoring in separate steps.

Patch-set: 2