Update patch set 3
Patch Set 3:

The only change between this patch and the previous is the deletion of the bug number in the comment's header. I'll create a second change to move the "mark" methods to JGit... then another to change the 3 places where these methods are copy/pasted to make use of this move.

Patch-set: 3