Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I tried this on Windows, but the problem didn't occur there.


So I simulated the problem by changing the code, and the workaround works well. Looks good to me.


Matthias: Time is running out for 2.3.1, should we submit this anyway?

Patch-set: 1