Update patch set 2
Patch Set 2:

Stefan,

Yes, some of the cases were already working. As you pointed out only two of them were not taking the ancestor into account.

The only reason I changed "all" was to factorize the code as much as possible since in the end the need is the same (and the code is 'almost' identical :).

Patch-set: 2