Update patch set 2
Patch Set 2:

I did not want to put in too much info in the commit message either :). I will add a paragraph to mention which were previously failing and which were only refactored.

If that's okay with you, I'll only push this change when the rest of the review is complete, that will avoid too many patch sets.

Patch-set: 2