Update patch set 2
Patch Set 2:

Matthias, Stefan,

While refactoring the "CommitFileDiffViewer" and testing in more details, I stumbled over a few things I would consider as bugs :

There are four actions that can trigger a comparison with a file that is not in the workspace; the last four I mentionned in the commit message.

The first two are failing with the current implementation because of a regression introduced by commit 5246f10 2 years ago. See this line exactly ( http://git.eclipse.org/c/egit/egit.git/tree/org.eclipse.egit.ui/src/org/eclipse/egit/ui/internal/history/HistoryPageInput.java#n74 ) that sets the input of the history page to a Boolean instead of a File. I'm fixing this with the next push for this review.

However, even when I fix that issue, a second one appears : the first of these two actions are comparing "HEAD" with the selected commit, while the last two compare "the local file" with the selected commit. I believe that the desired behavior is to compare the working directory, not HEAD, so I'll harmonize all four actions to have that behavior.

Unit tests would have caught that, but I am not familiar with SWTBot, so what I'll push here will only be the code, manually tested on my end.

Patch-set: 2