Synchronize - accurate revisions for remote and ancestor
This removes two unused (and outdated) tests in
GitResourceVariantTreeSubscriberTest. The third one was removed since
it was not testing what was advertised : it mocked a use case where the
user was synchronizing a branch with only one comit with a master that
also had a single commit, that thus acted as both the "remote" _and_
"common ancestor" of the synchronization. What we want is a use case
where all three revisions are distinct.

Bug: 393294
Change-Id: I69d4a53e23b7afe916a30b873aada13a6931f4c8