Update patch set 5
Patch Set 5:

I had a need for getTargets() and getContributors() ... and thus tried to test them without the "timing" restriction I previously had.

This new patch set reverts some of the changes made by the previous (namely, we use KidWalk anew now that I properly understand what getTargets() is supposed to do).

However, it also needs the opening of a new API: GitRemoteResource was package-protected, I changed it to public. Is that acceptable? Should it remain in its current package (exported as API), or should I move it to an internal namespace?

Patch-set: 5