Update patch set 2
Patch Set 2:

When switching branch Eclipse (EGit) leaves the derived objects around and then they get in the way during rebase. I need to revisit to be able to give you the details, but I discovered the problem when rebasing the symlink branch onto master and I couldn't do it without the command line.

Patch-set: 2