[stagingView] Improve usability of "Replace with" on missing
Only provide one of the equivalent options for a file that has been
deleted. For this to work, we now distinguish between missing MISSING
and MISSING_AND_CHANGED (which has changes in the index). For the latter
we still have to provide both options.

See also I65601162f7ef506ea270933218601322fde1e929 where the same change
was done for modified files.

Change-Id: Id4ff449791e7a911809281bab63f9d3e3d17b5ed