Update patch set 2
Patch Set 2:

Currently using 3.7 for that workspace. The warning comes from the API tooling. You can also manually go to the source of WorkbenchContentsFinder and see that the class Javadoc includes @noinstantiate, which is also used in other Eclipse API and means that it should not be created via new, hence the warning.

Patch-set: 2