Use the source revision instead of local data for synchronizations
EGit allows for comparisons with three remote revisions (notably through
"synchronize with each other" with two branches selected in the
repository explorer). However, Team always uses the local data as the
"source" (or left) side of a comparison. This patch allows us to use the
actual source commit instead of the local file when needed.

Bug: 403363
Change-Id: I23e0f1e75b9c6d476f8ed5a24deac7cc7b64dd42