GitFileHistory should return the full history of its target.
When using GitFileHistoryProvider.getFileHistoryFor(IResource,...), a
user expects the full history of the target resource to be returned,
except if the SINGLE_REVISION or SINGLE_LINE_OF_DESCENT flags are
positioned.

The SINGLE_REVISION flag is properly handled, however when no flags are
set, EGit previously returned only the commits descending from HEAD,
without considering the history that exists on other branches or
references.

This adds specific tests to make sure of the behavior when either no
flag or the SINGLE_REVISION flag are set.

The history's "getTarget" and "getContributors" were rewritten to allow
for some of Team quirks: when using the synchronize view, Team will use
its own file revisions instead of what was originally fed to it. We thus
need to adapt those revisions to types known to us. A specific test
demonstrating this use case has been added.

Bug: 398982
Change-Id: I67720110a47048fb85074a1cbbcb3d4205618a67