Update patch set 2
Patch Set 2:

Yes, the failure is most likely due to this change, but try as I might, I have not been able to successfuly run the SWTBot tests locally. I planned to update this change request with the updated test, but for now I cannot fix it. There is no use reviewing this for now, sorry about that.

Patch-set: 2