Eliminate file handle leaks
Running the SWTBot test suite on a windows environment highlights an
insidious file handle leaks coming from the WindowCache used in
conjunction with JGit parsed repositories. ".pack" files are opened for
reading and thus locked for deletion until the repo is properly closed,
but EGit might not keep a reference towards this repository in its
cache, "forgetting" to close it when needed. This is mostly true for
sub-modules.

This patch also gets rid of other random errors, sometimes due to a
.project not properly removed...

Bug: 404121
Change-Id: I3a596fcabdb5225e9042e1ffae6b054c48d72d54