Ambiguous parameter naming
What GirFileRevision really expects is a path, not a file name. Most
usages of the constructors already passed in a file path... but the
parameter naming led to a number of clients to use a file name, which
creates FileRevisions with no resolvable path. This fixes all
ambiguous parameter names and usages.

Change-Id: Id597d427555f71a6b4f422acb06a1faf26c796bd