New Subscriber to create ActiveChangeSetManager
To create an ActiveChangeSetManager (useful for Mylyn connector or other
similar plug-ins), I created a new Subscriber. GitUpdateSubscriber
extends GitResourceVariantTreeSubscriber, and contains some needed
modifications to generate a valid change set manager.

With this Subscriber, the cache of a specific resource is synchronized
each time this resource changes. Moreover, the new created Git projects
are saved in cache too.

Change-Id: I76ac813ec93dd64c9ee7440b9e8eb0719148bad1
Signed-off-by: Joffrey Chambrin <joffrey.chambrin@airbus.com>