Update patch set 3
Patch Set 3:

Robin,

I don't really understand what you mean by "add test for these cases"?

The only code within EGit itself that depends on the URI returned by the file revisions is the editor input, and it is only used as a displayed String.

I realized the inconsistency from my end as I use this path to determine where the file is and resolve "other" URIs that are relative to it. Whether I get a file system absolute path or a workspace relative path does not really cause any issue... as long as the path is resolveable (which is not the case for git repository relative paths for example).

The javadoc of that method tells us the expected result is "the URI of the file to which this state is associated or <code>null</code> if the file does not have a URI." ... I believe that a file system absolute path is indeed the best we can use to answer that.

Patch-set: 3