Only include resources that are part of the data set in synchronize
Workspace synchronization is launched on a given repository (or "a
project of a given repository"). This repository may contain commits
on files that were not imported in the workspace. In such a case, the
GitContainerMapping corresponding to that commit shouldn't create
traversals for the files that are not in the workspace, as that would
fail later on in NPEs.

This improvement may also serve as a basis to allow a "synchronize
project" (or "synchronize folder") action instead of "synchronize
workspace"; i.e. restrict the scope of a synchronization to only part
of a repository.

Bug: 405772
Change-Id: I824a6a88d701b2787e5c972414b8fd40407e2c8e