Update patch set 6
Patch Set 6:

This fix is very bad: it makes the left side of the compare editor read-only, so that one can no longer type and Copy Right To Left is not available.

Please either fix this asap or revert this change until a solution is available. Currently, the nightly build is unusable for me due to this bug.

Patch-set: 6