DRAFT - factorize comparison and synchronization code
* Use the same code for all comparison actions, making sure that
a comparison launches a synchronization instead when there is more
than a single resource to compare (either when the user selected
more than one resource or when the selected resource is part of a
larger logical model).
* Remove all now useless code from the comparison action handlers,
CompareUtil and GitSynchronizeModel.
* Allow GitSynchronizeData to filter on a set of resources instead
of filtering on a set of containers. This finer-grained filtering
allows us to constrain the scope of synchronization operations.
* Remove useless overrides from GitModelTree and GitModelRepository.

Change-Id: I4ae8abfc4dbe7ed772749afa80743278247585ee