Only include resources that are part of the data set in synchronize
Workspace synchronization is launched on a given repository (or "a
project of a given repository"). This repository may contain commits
on files that were not imported in the workspace. In such a case, the
GitContainerMapping corresponding to that commit shouldn't create
traversals for the files that are not in the workspace.

Bug: 405772
Change-Id: I824a6a88d701b2787e5c972414b8fd40407e2c8e