Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I also can't reproduce the problem anymore, probably some other changes have changed the behavior of the staging view, will need to dig into that.

Also there seem to be some strange selection problems, it seems the effect of the selection change event on the state of the toolbar buttons depends on which views were selected before clicking on a staging entry. This worked earlier.

Patch-set: 2