Fixed bug #406922: stronger checks for handling linked resources.
The fix provides stronger checks on selected linked resources before
enabling menu items. In particular, linked resources that point to a
file that is part of a project shared with egit are allowed to be
operated on. Otherwise team actions are disabled if the selection
contains a linked resource pointing to a file outside any project and
repository. A default isEnabled() implementation is provided for this
behavior. After adding a test case a couple more handlers have been
amended and made compliant.

Bug: 406922
Change-Id: Id9fd5ea0c31a78b84796638570d214d14cc3c4fd
Signed-off-by: François Rey <eclipse.org@francois.rey.name>