Declare everything in EGit internal
What appears to be API's in EGit is not well designed and arbitrary.
Basically we forgot to put 'internal' into the package naames.

This change doesn't break current users, but warns that the API's are
not stable and subject to future improvement. Users that want to use the
(now internal) API's in upcoming major revisions should ask for this
explicitly.

Change-Id: I0f0e4186cc818532e5c4d7c31cff877fb3591561