Prefer to create new branch based on remote tracking branch
Prefer to use a remote tracking branch as the base of a new branch if it
is created by selecting a commit in the history view multiple branches
are referring to. This makes sense since most often the user wants the
new local branch to track a remote tracking branch.

Change-Id: Ie1e863a5fb3d274dd215f1531f3da5cc43b08de5
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>