Update patch set 1
Patch Set 1: (6 inline comments)

Displaying the filter field as shown in the mockup is challenging, particularly right aligning it with Unstaged Changes and Staged Changes sections, and handling the case where the vertical sash is moved, the selected repository changes, etc.  I have made the change to show "Filter files" as the search message rather than "Find".  Maybe this makes it more obvious what the control's purpose is, even if it is in its current location?

I've fixed the unstaged/staged panes to show <shown files>/<total files> in the section text, as suggested.

When I revisit the other change, adding the display options, I'll be sure to make sure that it behaves correctly with respect to filtering.

Patch-set: 1