Update patch set 1
Patch Set 1: (1 inline comment)

I don't believe my patch is the true underlying cause of the NPE.  Without the patch, you will get a similar NPE if you manually select the same file that gets selected automatically with the patch.  The error is triggered by the selection, whether it is manual or automatic.

I don't mean to quibble over details and, if you'd like, I can try to figure out where the problem was introduced and fix it.  I'm just pointing this out because I'm not sure that you wouldn't prefer to file a separate bug for this.

Patch-set: 1