Update patch set 1
Patch Set 1: Verified; Looks good to me, approved; Unclean IP, do not check in

we need you to answer the 3 contributor questions as a comment on this review or (hurray) you may file a CLA using the new process described here [1], as your change is already signed off filing the CLA is equivalent to the old process. With the CLA in place you don't need to answer the 3 questions anymore but only sign off each change you contribute.

[1] http://mmilinkov.wordpress.com/2013/06/17/eclipse-clas-are-live/

Patch-set: 1