[stagingView] Add presentation options (list, tree, compact tree)
This adds different options for presenting the unstaged and staged files
in the staging view.

"List" is the same as before, just a flat list of files, e.g.

  file1 - src/org/eclipse
  file2 - src/org/eclipse

"Tree" shows the files in a full folder tree, e.g.

  * src
    * org
      * eclipse
        * file1
        * file2

"Compact Tree" shows the files in a folder tree where folders that have
just one subfolder do not result in a node, e.g.

  * src/org/eclipse
    * file1
    * file2

When an expanded node is dragged from one tree to another
(stage/unstage), the expanded state should be preserved.

CQ: 7377
Bug: 407607
Also-by: Robin Stocker <robin@nibor.org>
Change-Id: I6509f2ee0f63fccf0391418cc3d41223e71c41de
Signed-off-by: Stephen Elsemore <selsemore@collab.net>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>
Signed-off-by: Robin Stocker <robin@nibor.org>