Fix scheduling rule error when ignoring resources in nested projects
The problem was that IgnoreOperation#calcSchedulingRule and
IgnoreOperation#addIgnore used different methods to get an IContainer
from the IPath. In case of nested projects, this could result in them
getting different IContainers, which lead to an error like "Attempted to
beginRule: /container1, does not match outer scope rule /container2".

To make the test case check the rule, it had to be wrapped in a job.

Bug: 410011
Change-Id: Id03daa1db356b8dbd0b4ad1b07173760b3e1bc75
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>