GitSyncObjectCache erroneously changes the sync state of resources
When using the mapping context to determine the synchronization state
of individual resources, the GitSyncObjectCache changes all other
resources' state to "in-sync"... and never looks back. This causes
very weird issues when using Team APIs, and a totally unspecified
behavior such as a resource going from "MODIFIED" to "IN-SYNC" for no
obvious reason.

Bug: 415430
Change-Id: Ib9d3686e87a5cb490a55b3a379b81b0fc0b1c3d9
Also-by: Laurent Goubet <laurent.goubet@obeo.fr>
Signed-off-by: Mikael Barbero <mikael.barbero@obeo.fr>