Fix instability in MergeToolTest
It recently failed here:
https://hudson.eclipse.org/sandbox/job/egit.gerrit/4746/

The problem was that the "ancestor" text editor was sometimes shown and
returned as the first text editor, where the code expected to get the
"left" editor.

As the sync view tests also use the compare editor, abstract this in a
new CompareEditorTester which handles this.

Change-Id: I82e731f4166c5cf060a8f441b02a6067961630bf
Signed-off-by: Robin Stocker <robin@nibor.org>