[repoView] Make "Delete Repository..." more explicit
It now requires the user to explicitly check that the .git dir should be
deleted before the dialog can be confirmed. It also displays a warning
that the deletion is permanent and changes some labels for clarity.

The paths are now shown below the checkbox for better alignment and to
make them selectable (see "Delete Project" where this is also done).

Bug: 395351
Change-Id: Ia1b95ecc1244e4dd622a0e6b77f665d13327ac80
Signed-off-by: Robin Stocker <robin@nibor.org>