Update patch set 1
Patch Set 1:

The idea is that deleting .git has to be a conscious action from the user (by selecting the checkbox). When it's not selected, the action can not be executed at all (because nothing would be done). Do you think we should only allow the user to select it, but not deselect it?

About the grammar of the message, it was like this before this change, but I can change it too.

Patch-set: 1
Label: Verified=0