Update patch set 3
Patch Set 3:

> - I could want to just remove projects from the workspace

That is what "Remove from view" is for, it asks whether to also remove the projects. I don't know why anyone would search for this functionality in a command named "Delete Repository...". It would have to be renamed to "Delete Repository or Maybe Not" ;).

> - I could want to delete the working tree without deleting the repository

I really don't see the point of doing that. You would end up with a repository that has all its files marked for deletion (if the deletions are automatically staged like they are when deleting files e.g. from the navigator). Could you make an example where this would be useful?

> - I could want to delete the repository without touching working tree and imported projects

That's possible in the original dialog and the proposed modification.

I guess my main gripe with enabling it when anything is selected is that the action may no longer really be a "Delete Repository". I would be pretty hesitant to click on the "OK" button of a dialog titled "Delete Repository" when I'm not expecting it to delete anything.

Patch-set: 3
Reviewer: Gerrit User 79 <79@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0