Update patch set 5
Patch Set 5:

Laurent is currently in vacation. I would have liked to address the comment, but I cannot push a review on behalf of Laurent. If I change the "signed-off", the commit review process tells me that the author of the commit has not "signed-off" the commit. 

Any idea if it could be done before Laurent come back?

Patch-set: 5
Label: Verified=0