Update patch set 6
Patch Set 6:

The solution was to reference Laurent in "Also-by" and to change the author of the contribution to my name. I hope Laurent and you does not bother about that. I did it to speed up the integration of this patch.

Patch-set: 6
Reviewer: Gerrit User 1007 <1007@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0