Update patch set 6
Patch Set 6: Code-Review-1

I think I can change it before submitting it, so we don't have to wait for the verification to be changed.

I just noticed that shouldBeIncluded also exists in GitResourceVariantTreeSubscriber and has nearly the same implementation (but it doesn't check for null return of getLocation). Could we move the implementation into a method on GitSynchronizeDataSet instead, so the code is shared?

Patch-set: 6
Label: Code-Review=-1