Update patch set 3
Patch Set 3: Code-Review-2

The fix is not correct and makes the test green only by accident. "key" in GitSyncCache is the Git dir path of the repository, while updateRequests contains file paths.

The fix must pass updateRequests to GitSyncObjectCache#merge and it must figure out what it may update and what not. Note that it's recursive and we probably also have to pass along the parent path.

Try the following test. It's not yet complete as we also need to e.g. test what happens when a new file in a new directory is added, but it's a start:

	@Test
	public void hasLocalChange() throws Exception {
		File file1 = testRepo.createFile(iProject, "a.txt");
		File file2 = testRepo.createFile(iProject, "b.txt");

		testRepo.appendContentAndCommit(iProject, file1, "content a", "commit a");
		testRepo.appendContentAndCommit(iProject, file2, "content b", "commit b");

		JGitTestUtil.write(file1, "changed content a");
		JGitTestUtil.write(file2, "changed content b");

		IFile iFile1 = testRepo.getIFile(iProject, file1);
		IFile iFile2 = testRepo.getIFile(iProject, file2);

		RemoteResourceMappingContext context = prepareContext(MASTER, MASTER);
		assertTrue(context.hasLocalChange(iFile1, new NullProgressMonitor()));
		assertTrue(context.hasLocalChange(iFile2, new NullProgressMonitor()));

		JGitTestUtil.write(file2, "content b");
		refresh(context, iFile2);
		assertTrue(context.hasLocalChange(iFile1, new NullProgressMonitor()));
		assertFalse(context.hasLocalChange(iFile2, new NullProgressMonitor()));
	}

	private void refresh(RemoteResourceMappingContext context, IFile file)
			throws CoreException {
		context.refresh(new ResourceTraversal[] { new ResourceTraversal(
				new IResource[] { file }, 1, 0) }, 0, new NullProgressMonitor());
	}

Patch-set: 3
Label: Code-Review=-2