Update patch set 4
Patch Set 4:

Thank you for your review and insights and sorry about the monkey patch. I've added your test along with other with sub-folders. It works pretty well. I still have an issue with hasRemoteChanges. Here is a symbolic representation of the commit

* |   C
| *   B
|/ 
*     A

First I created a file with a content "X" and commit it in commit "A". I modify it to content "Y" in commit "B" and re-checkout commit "A". So far, everything is as expected, i.e. context has remote change and no local change. I change back the content of the file to "X". Here is the issue, the context has no local change but I would expect it not to have remote change neither. What do you think? The same behavior occurs if I commit the change to commit C, the context has remote change and no local.

If you agree with my expectations, the issue seems to come form org.eclipse.egit.core.synchronize.ThreeWayDiffEntry.scan(TreeWalk). During the walk, the local id of the file is the zero id:

walk.getObjectId(idBuf, 0);
e.localId = AbbreviatedObjectId.fromObjectId(idBuf);

Where it should be the same as the remote. Any idea how it should handle pseudo conflict like this?

Patch-set: 4
Label: Verified=0