Update patch set 4
Patch Set 4:

First, I'm in the following state

  | *   B <-- BRANCH, file contents = "state 2"   
  |/ 
  *     A <-- MASTER | HEAD, file contents = "state 1"

I'm comparing MASTER to BRANCH. The context properly reports that it has remote change but no local one. Then, in branch MASTER, I set the file contents to "state 2" and commit it. The file is exactly the same on BRANCH and MASTER.

  * |   C <-- MASTER | HEAD, file contents = "state 2"
  | *   B <-- BRANCH, file contents = "state 2"   
  |/ 
  *     A 

The problem is that the context is still reporting it has remote change but no local change. I would expect either no change, neither locally nor remotely, or that it has changes on both sides.

Patch-set: 4