Update patch set 2
Patch Set 2: Code-Review-2

(1 comment)

I don't quite get how the selection dialog is supposed to work in "Fetch a change from Gerrit".

When I select an existing branch with "Select" and then click "OK", an error "A branch or tag with name foo already exists" is shown. So apparently I am supposed to enter a name of a new branch there. In that case, it doesn't make sense to name the button "Select".

Is it e.g. so that the user can first delete an existing branch to make the name available again?

In that case, it should probably be named something like "Edit Branches..." and "Cancel" and "OK" should be replaced by "Close" and not do anything - it currently writes the name of the selected branch in the text field, which triggers the error above.

Also, "New Branch..." doesn't make much sense in that case, or does it? Do you have an example use case for doing this in that context?

Patch-set: 2
Label: Code-Review=-2