Update patch set 2
Patch Set 2:

Since we don't need a release review for a maintenance release we can accept non-committer changes until we create the 3.0.2 release. 

Though the size of this change is > 250 LOC so it looks like we need to file a CQ for it. Which fraction of this change comes from you and how much from Laurent and Mikael?

Patch-set: 2
Label: IP-Clean=0