[sync] Fix addition/deletion of incoming commit being reversed
When a file is added as part of an incoming commit, it should not be
shown as deleted and vice versa. I don't know why this was done in the
first place.

There was also confusion about what the LEFT and RIGHT direction mean in
the context of synchronization. The static block in
org.eclipse.compare.CompareConfiguration shows that LEFT is incoming and
RIGHT is outgoing.

Bug: 416374
Change-Id: I4b68e2b35b50fe207180e4313868c1bcfecd87c2
Signed-off-by: Robin Stocker <robin@nibor.org>