Update patch set 5
Patch Set 5:

.equalWidth(true) caused the buttons to be clipped when the view is small

So I changed the layout to use 2 cells for the commit&push button since it has a much wider label and gave up on equal button width. With that the clipping only occurs when the view gets smaller than the width needed for the buttons.

Any SWT wizard who has a better proposal for this layout problem ?

Patch-set: 5
Label: Verified=0