Update patch set 3
Patch Set 3:

> I think fetch is used more often than push if we assume the average team has more than 2 developers and developers push a similar number of commits per day.

It depends on the workflow. In non-Gerrit projects with a central repo, pull is used most often, followed by push, then fetch.

> This would mean we should rather reverse the order in git menu and repositories view to position most frequently used commands (fetch) at the top of the menu in order to minimize mouse movements. 

If the menu opens upward (when the repo view is positioned at the bottom), then fetch at the top would require the most mouse movement. I personally use the toolbar, which is probably the quickest way.

Would you also like to change the order of the toolbar buttons (currently push, fetch, pull)?

What makes the whole thing a bit more complicated is that "Pull" in the Team menu is in its own section below fetch/push. So changing the order to fetch, pull, push would require moving pull around there. Or would you want to make the order fetch, push, pull? That seems kind of weird, as fetch and pull are related and so should be next to each other.

To be honest, I don't mind much as long as the order is consistent in the menus. What should we do to not have this stalled (I would like to get "Push Branch..." into 3.1)?

Patch-set: 3
Reviewer: Gerrit User 79 <79@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0