Update patch set 6
Patch Set 6:

> I feel it's not obvious what is the source branch and what is the target branch of the push being configured here

I'm not sure what you mean by source and target.

The branch being pushed is either the selected branch or the current branch if invoked via team menu. It is shown in the wizard title (e.g. "Push Branch foo").

What the branch should be named on the remote (which is usually the same as the local name) can be entered on the page in the "Branch name" text field.

> how can I configure the upstream branch I want to push to ?

Can you give an example of what you are trying to do?

The wizard does the same as `git push --set-upstream remote-name` when the current branch is selected. If another branch is selected, it's `git push --set-upstream remote-name branch-name`. If the branch name is changed in the text field, it's `git push --set-upstream remote-name local-branch-name:remote-branch-name`.

Patch-set: 6