Update patch set 7
Patch Set 7: Code-Review-1

(4 comments)

Nice that we finally get this, see inline comments.

Regarding the UI, has it been considered to place the rebase buttons in its own "Rebase" sectuion, and put it *above* "Commit Message"?

One argument for that placement would be that the flow of the user interaction should be from top to bottom. In case of a typical rebase, what the user does after resolving conflicts is click on "Continue", and the commit message and commit buttons are not used. With the current placement, the commit message and commit buttons have to be skipped visually.

An argument against that is that in the case of a merge with conflicts, the button to click at the end is also at the bottom right, "Commit".

So I'm not sure what the best thing is here. Opinions?

If we leave the placement below commit message, I think we should change the section title from "Commit Message" to something else or put the rebase buttons in a "Rebase" section (but leave them at the bottom).

Patch-set: 7
Label: Code-Review=-1