Add class that represents the rebase-todo and rebase-done files
The RebaseInteractivePlan provides a list that joins both the todo and
done file. Elements can be reordered and changed. Any changes to the
todo-elements are immediately written to the todo-file to avoid any
difference between todo-elements in the RebaseInteractivePlan and the
rebase-todo-file.

The RebaseInteractivePlan is registered as IndexDiffChangeListener.
This guarantees that the plan is notified on all relevant actions of the
interactive rebase and can refresh itself from the file system.

JGit-Dependency: I323f3619952fecdf28ddf50139a88e0bea34f5ba
Change-Id: Ie698bb77f9095ffb8184e50c47288ff8ee8eb0ae
Signed-off-by: Tobias Pfeifer <to.pfeifer@web.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>
Signed-off-by: Stefan Lay <stefan.lay@sap.com>