Add class that represents the rebase-todo and rebase-done files
The RebaseInteractivePlan provides a list that joins both the todo and
done file. Elements can be reordered and changed. Any changes to the
todo-elements are immediately written to the todo-file to avoid any
difference between todo-elements in the RebaseInteractivePlan and the
rebase-todo-file.

JGit-Dependency: I323f3619952fecdf28ddf50139a88e0bea34f5ba
Change-Id: I2cc82878a265e3ec8bb7f8dca90011f08f3c0aff
Signed-off-by: Tobias Pfeifer <to.pfeifer@web.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>