Add class that represents the rebase-todo-file and rebase-done-file.
The RebaseInteractivePlan provides a list that joins both the todo and
done file. Elements can be reordered an changed. Any changes to the
todo-elements are immediately written to the todo-file to avoid a gap
between todo-elements in the RebaseInteractivePlan and the
rebase-todo-file.

Depends on JGit change: I323f3619952fecdf28ddf50139a88e0bea34f5ba

Change-Id: Ie698bb77f9095ffb8184e50c47288ff8ee8eb0ae
Signed-off-by: Tobias Pfeifer <to.pfeifer@sap.com>