[historyView] Include annotated tags with "Show All Branches and Tags"
When a commit was only reachable through an annotated tag, it was not
shown in the history view even when using the show all option.

The problem was that parseAny returns a RevTag for annotated tags, not a
RevCommit. By peeling the returned object, we get the commit the tag
points to and include it in the walk.

Bug: 417655
Change-Id: If1b64f9eb56c49057598486a455d5ccd08928836
Signed-off-by: Robin Stocker <robin@nibor.org>