Make the rebase command stop after initialization
Preparing the rebase todo lines in a callback method may be a long
lasting user interaction. Doing this with the callback method
prepareSteps(...) in the RebaseInteractiveHandler implies to make the
thread that is calling back sleep until the user has finished
preparation. When the rebase command is returning after initialization
there is no need to make a thread sleep. The rebase todo file can be
edited via writeRebaseTodoFile(...) of the Repository class. The rebase
(interactive) has then to be invoked with PROCESS_STEPS in order to
start the processing.  Note that the Method prepareSteps(...) of the
RebaseInteractiveHandler will be called anyway just before starting to
process the rebase todo file.

Depends on JGit Change-Id: Change-Id:
I2d6f0de82010ea6523fbce6fb4501e847bdcdddc

Change-Id: I52a1fdae82308f3c7e4e4950b007e83807d0ae41
Signed-off-by: Tobias Pfeifer <to.pfeifer@sap.com>