[stagingView] Start with "auto expand" mode initially
As a default, I think having the tree expanded makes more sense than
collapsed, so that the files can be seen instead of just some parent
folders. Using the compact tree, I caught myself always pressing "Expand
All" as the first step after starting Eclipse, and I've seen other
people being bothered by having to expand the nodes (I don't know why
they didn't use the "Expand All" button though).

In the list view, we were always showing all files, so performance
should not be a problem. And as soon as the user collapses a node, "auto
expand" is disabled as implemented in an earlier change.

In the future, we could also consider persisting the state.

Change-Id: Iac2f9278a25efdee7e25991612dc7ba8cbcc0648
Signed-off-by: Robin Stocker <robin@nibor.org>