Improve UI of upstream configuration when creating a branch
The bare "Pull strategy" group with "Rebase", "Merge" and "None" as
radio buttons works, but requires much prior knowledge of what it does.

This change replaces it by a more descriptive set of controls, and makes
it clear that the upstream configuration is also used for pushing. It is
provided as a separate component for reuse in the "Push Branch" wizard.

Change-Id: I481140cefd53198c8b9bce5adbb77d52dc34fa80
Signed-off-by: Robin Stocker <robin@nibor.org>