Update patch set 7
Patch Set 7:

> Nitpicking: once I created a new remote in the wizard, I'm no longer able to select the other ones. Other than that looks good to me.

This is deliberate because the remote is only created if the wizard is finished, but not if the user cancels the wizard. So if the user would select "New Remote..." but then select a different one, they might expect the remote to be created. Therefore the selection is locked. It also keeps the implementation simpler :).

Patch-set: 7
Label: Verified=0