Update patch set 1
Patch Set 1:

I see, the problem is that the rebase hasn't even been started yet, and therefore aborting it only changes the repository state, but not the index diff nor refs.

Hm, I'm not sure what to do. I don't like firing refschanged for all commands, as it results in unnecessary work in IndexDiffCacheEntry.

Patch-set: 1