Fix UI tests interfering with each other because of shared test setup
The most common cause of unstable UI tests was that the test data
(repositories and projects) was set up once per class, and then each
test method used the same data. When one test method only reads the data
but another modifies it, the test may fail depending on the order of
execution.

This is an attempt to fix this by changing the test setup so that it's
done once per test method (@Before instead of @BeforeClass). To make it
easier (and possibly faster) to clean up and reduce the possibility of
accidentally sharing the data, the data for each test method is created
in a separate directory.

Change-Id: I2d7ae4e86c08483a1e8a2824a31648f9d6827bf8
Signed-off-by: Robin Stocker <robin@nibor.org>