Fix ReplaceActionsTest instability caused by commit time stamps
It failed when the time between the two commits happened to not be in
the same second (Git commit time stamp has second resolution).

To fix it, make sure the time stamps are always different and change the
order in the dialog to correspond to the order in the history (show
newer commits first).

Change-Id: Ie7e992782cff55b65c7297cd91dbd43891cb6501
Signed-off-by: Robin Stocker <robin@nibor.org>