Fix misleading error message in clone wizard
When a remote repository could not be reached there was a dialog which
only described some of the possible scenarios. Moreover, it always
proposed to look at the ssh preferences, even if ssh was not used.

We now give more explanation about possible failures. Possible errors
in case of ssh or https are only shown when these protocols are used.

The Exception was not logged, this is fixed now as well.

In further change we could also react on the concrete type of the
Exception thrown inside JGit. E.g., an
org.eclipse.jgit.errors.NoRemoteRepositoryException would indicate that
there is no network problem but a wrong repository path.

Bug: 420703
Change-Id: Ia162aa0485496b8e4306a941ad1318a5db4422f2