Improve rendering of file diffs in History view
The Comment pane in the History view should render a blank line after
the commit message and before the diffs for the files selected on the
right.

The fix for bug 410220 made this old rendering bug much more apparent,
especially in repos with terse commit messages.

This fix also renders EGit's additional
------ path/to/file.ext ------
header line in grey to make it stand out from the "diff -git ...".

Change-Id: I4ed779e62c1ea2d58e6d84cc6271743b928dcc10
Also-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Markus Keller <markus_keller@ch.ibm.com>