Update patch set 3
Patch Set 3:

By the way, I just saw that the bug report did not have a comment pointing to this review.

When you pushed a change which fixes a bug, we normally change the bug to ASSIGNED and add a comment pointing to the review, so that other people coming to the bug can find it.

Patch-set: 3