Blame annotations: Show diff in hover and enable blaming parent commits
This change extends the information that is shown when hovering over a
annotation revision:

A diff to each parent (normally just one) is shown. Only the hunk
concerning the current line is used to keep the diff small. This helps
in understanding what the blamed commit changed in the current line.

The text above the diff shows the parent commit ID and includes a link
to show the annotations of this parent commit. When clicked, a new
editor is opened for showing annotations for that commit. It also jumps
to the correct line.

Together, this enables quickly going through the history of a file and
inspect the changes that happened over time.

Bug: 421909
Change-Id: I11793926570a057577c6de04c3a8bda28521d71d
Signed-off-by: Robin Stocker <robin@nibor.org>