Blame annotations: Show start commit when opening editor
Before, the commit that was shown was the blamed commit from the cursor
position, which could be confusing to the user.

To solve this, two things needed to be changed:

The first time a selection changed event happens for the revision
selection provider, we ignore it as it is for the "initial" selection.
The confusing thing about this is that the user has not yet consciously
selected a revision. To fix this, we now always ignore the very first
selection.

When opening an editor with annotation for a start commit (e.g. for the
show annotations link in a hover), we now show this commit in the
history view.

Bug: 422100
Change-Id: I5e1e529684f298b51fea5ccf99f18d901b874c5a
Signed-off-by: Robin Stocker <robin@nibor.org>