Push Tags wizard
Can be started with one or more selected tags from the repositories
view. It's also available from the team menu and the context menu of a
repository.

In the "Create Tag" dialog, there is an additional button which can be
used to create the tag and then start the push wizard for it.

Bug: 341076
Change-Id: Idbc4b6e241c37548ffa2e9697697ad3965de08e7
Signed-off-by: Robin Stocker <robin@nibor.org>