Do not allow rebase if there are uncommitted changes
With this change egit checks for uncommitted changes before a rebase is
started. The user can choose from a dialog how to handle this situation.
In a further change we could offer the feature to automatically re-apply
the stashed change after finishing or aborting the rebase.

Note that this change is done in EGit. Therefore the Pull operation is
not affected.

Bug: 422352
Change-Id: I7e7878daa67d3316863b6ebbc40f61fad0460aa0
Signed-off-by: Stefan Lay <stefan.lay@sap.com>