Rebase: Show dialog in case there are uncommitted changes
JGit now checks if there are uncommitted changes before the rebase is
started. In such a case the user can now choose from a dialog how to
handle this situation.
In a further change we could offer the feature to automatically re-apply
the stashed change after finishing or aborting the rebase.

Bug: 422352
Change-Id: I7e7878daa67d3316863b6ebbc40f61fad0460aa0
Signed-off-by: Stefan Lay <stefan.lay@sap.com>